% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dggapm.R
\name{dggapm}
\alias{dggapm}
\title{dggapm}
\usage{
dggapm(A, p, h)
}
\arguments{
\item{A}{gene expression matrix of the studied genes}

\item{p}{number of genes involved in the calculation}

\item{h}{a parameter}
}
\value{
a matrix representing pairwise precison mattrix and a binary adjacency matrix indicating the presence of network edges between genes
}
\description{
Calculate high-dimensional precision matrix for a given set of genes and return the partial correlation matrix, with edges of the gene network estimated through adaptive thresholding
}
